Send flavor resources as value metadata
This addresses an issue where the resources dimension could
contain a string with prohibited characters, for example:

"VCPU:64, MEMORY_MB:393216, DISK_GB:110, None".

This change moves the flavor resource string to the value_meta
field which has fewer restrictions over the content. If we
later want to index resources as a function of the flavor
resource fields, for example CPU count, we could add those
as separate dimensions on the metric.