Create host aggregate relation metric
Create openstack_hypervisor_placement_aggregate_relation
metric which contains hypervisor name,
and list of aggregates that a hypervisor is part of,
in addition to the availability zone. This metric will be
used later with other promql commands to determine the
usage of a compute node within it’s associated aggregate.

Jira: IAAS-31910